* Moved things around and organized them by the part of the file they work with: Header, Entries, Entry
* Simplified some things by storing a copy of the Entries object
in Header so it can figure out what to put in Line-Count. This is
a tiny bit distasteful, but I don't want to do the work to make
some thing that watches all of the objects.

* The Entries object moves around a little too. It stores the
column names so it doesn't have to ask the Header object for them.

* I added a first pass at mapping external and internal Header field
names. Externally (in the output), they have initial caps on the words
and hyphens. Internally, I want those to be Perl identifiers, so I
lowercase them and change the hyphens to underscores. There is a little
bit of duplicated effort that I need to clean up, and the docs on this
need more coherence.

* All of the tests pass, and the coverage is okay:

---------------------------- ------ ------ ------ ------ ------ ------ ------
File                           stmt   bran   cond    sub    pod   time  total
---------------------------- ------ ------ ------ ------ ------ ------ ------
...ib/CPAN/PackageDetails.pm   95.3   78.6   73.3   98.0   90.0  100.0   92.8
Total                          95.3   78.6   73.3   98.0   90.0  100.0   92.8
---------------------------- ------ ------ ------ ------ ------ ------ ------