* Improve _filter_old_dists by watching for order of additions
I was pretty stupid with the previous way I did this. I tracked
everything in %Seen, put returned the list to the next map anyway.
If the older dist showed up in the list first, I passed it on because
I hadn't seen anything newer yet. The list order matters.

I don't completely like how I've done it now, but it works.